mylibrary

+++ src/main/java/com/library/controller/LendController.java
@@ -30,10 +30,15 @@ public class LendController {
     @RequestMapping("/deletebook.html")
     public String deleteBook(HttpServletRequest request, RedirectAttributes redirectAttributes) {
         long bookId = Long.parseLong(request.getParameter("bookId"));
-        if (bookService.deleteBook(bookId)) {
-            redirectAttributes.addFlashAttribute("succ", "图书删除成功！");
-        } else {
-            redirectAttributes.addFlashAttribute("error", "图书删除失败！");
+        if(bookService.bookisLend(bookId)){
+            redirectAttributes.addFlashAttribute("error", "图书被借阅，不能删除！");
+        }
+        else {
+            if (bookService.deleteBook(bookId)) {
+                redirectAttributes.addFlashAttribute("succ", "图书删除成功！");
+            } else {
+                redirectAttributes.addFlashAttribute("error", "图书删除失败！");
+            }
         }
         return "redirect:/admin_books.html";
     }
